+++ src/containerapp/azext_containerapp/commands.py
@@ -55,6 +55,12 @@ def load_command_table(self, _):
         g.custom_command('remove', 'remove_env_managed_identity', supports_no_wait=True, exception_handler=ex_handler_factory())
         g.custom_show_command('show', 'show_env_managed_identity')
 
+    with self.command_group('containerapp env storage') as g:
+        g.custom_show_command('show', 'show_storage')
+        g.custom_command('list', 'list_storage')
+        g.custom_command('set', 'create_or_update_storage', supports_no_wait=True, exception_handler=ex_handler_factory())
+        g.custom_command('remove', 'remove_storage', confirmation=True, exception_handler=ex_handler_factory())
+
     with self.command_group('containerapp service', deprecate_info=self.deprecate(redirect='containerapp add-on', expiration='2.59.0', hide=True), is_preview=True) as g:
         g.custom_command('list', 'list_all_services')