+++ src/elastic-san/azext_elastic_san/aaz/latest/elastic_san/_update.py
@@ -13,7 +13,6 @@ from azure.cli.core.aaz import *
 
 @register_command(
     "elastic-san update",
-    is_preview=True,
 )
 class Update(AAZCommand):
     """Update an Elastic SAN.