+++ src/containerapp/HISTORY.rst
@@ -12,6 +12,8 @@ upcoming
 * 'az containerapp create/update': support --max-inactive-revisions
 * 'az containerapp env create': support --mi-system-assigned and --mi-user-assigned for environment create commands
 * 'az containerapp env identity': support for container app environment assign/remove/show commands
+* 'az containerapp env storage set': Support create or update managed environment storage with NFS Azure File.
+* 'az containerapp up': Update the Docker error string used to identify unauthorized push.
 * 'az containerapp env java-component': support list Java components
 * 'az containerapp env java-component spring-cloud-config': support create/update/show/delete Spring Cloud Config
 * 'az containerapp env java-component spring-cloud-eureka': support create/update/show/delete Spring Cloud Eureka